Expanded intermediate (IR) form for R, adding code to evaluate the IR in interpreter mode. Next step is generating byte code!
For an R expression that looks like this:

y<-1
for(i in 2:4) {
  y <- y * i
}
y

Has IR that looks like:

     0:  y := 1.0
     1:  _t2 := primitive<:>(2.0, 4.0)
     2:  _t0 := 0
     3:  _t1 := primitive<length>(_t2)
     4:  goto L1
L0   5:  i := _t2[_t0]
     6:  y := primitive<*>(y, i)
     7:  increment counter _t0
L1   8:  if not _t0 >= _t1 goto L0
     9:  return y


The most interesting instructions are dynamic<> and primitive<>.
Dynamic<> instructions invoke a function through the normal R pathway:
- looking up a symbol
- (possibly) dispatching to a class-specific function depending on the class attribute
- named argument matching

Tne goal is to resolve as many dynamic<> calls to specific primitive<> calls.
We (often) have specific type information about primitive functions, so we 
can then take some short cuts by not converting the variable 'y' back and forth
to a vector because we should be able to deduce that it actually can be 
stored as an 'int' throughout.







git-svn-id: http://renjin.googlecode.com/svn/trunk@421 b722d886-ae41-2056-5ac6-6db9963f4c99